Use auto-mode-alist to automatically set modes based on filename
Do this for Makefiles and Python *.py files.  If we open a Makefile
that is named GNUmakefile, makefile, or Makefile then set makefile-mode.
The main thing here is that tabs are enabled.  For *.py files, set
python-mode.

I am looking at other modes and file types that I commonly work with
and figuring out what I want.