Correctly start gpg-agent from .zshrc
With GnuPG 2.x the agent stuff is much easier to work with.  This
commit adds a block which sets the variables required and starts the
agent to ensure ssh works since I will probably use that before gpg in
a session.

If it's already running, these commands are a no-op and it just
connects to the running agent.  That should make sure things continue
working if the X session, for instance, starts gpg-agent.