Add .bash_profile and .bashrc to exec over to zsh.
On systems that prevent chsh or force the use of bash, have my
dotfiles exec zsh on startup of bash as if nothing was out of the
ordinary.

Login shells will logout on Ctrl+D as expected and interactive shells
will return to the calling program on Ctrl+D as expected.