Add xbindkeys and evrouter config files and start daemons in .xsession
This adds my .xsession script which is used on *most* systems I log in
to.  I start xfce4 but before that I do some things that xfce4 doesn't
handle itself.  This now has a block that starts xbindkeys and
evrouter on my workstation that has a Griffin PowerMate USB control
knob.  This is used to mute and raise/lower audio volume.