Finally fix the annoying inconsistent Home/End keys.
These were behaving differently on various systems at the shell, in
emacs, and inside tmux.  I didn't know what was at fault, so I backed
everything down to the shell init file.  It has two bindkey lines.  I
set Home and End, but that code was not consistent across all terminal
types.

So, now .zshrc will set Home and End to be the terminal definition's
Home and End key.  This requires tput(1) and that you have correct
terminal settings, but otherwise it seems to finally solve the problem
across all systems and inside emacs and tmux.