Start a ~/.git_template tree with common hooks
Now when I clone a repo or run "git init" in an existing repo, I will
get these hook scripts.  Here's what I have so far:

1) ctags which is called from post-checkout, post-commit, post-merge,
and post-rewrite.  This keeps the ctags file up to date.  I use emacs,
so this runs "ctags -e" and the file is called TAGS.  For vim it would
be just "ctags" and the tags file is called "tags".

2) The pre-push script catches !fixup and !squash commits preventing
you from pushing them to a remote.  Allows you a chance to fix things
up before pushing.