Better HTML email viewing in mutt, use geeqie for images.
Better command line for html2text for HTML formatted emails.  This is
primarily to deal with embedded links in things from Jira where it
will give a long autogenerated URL that you need to click.  This
command line wraps links in < > and prevents them from wrapping the
screen, which is useful if you need to click it or copy/paste it.

For now stick with geeqie for image viewing.  I am still working on a
way to inline view images using copiousoutput, even if it is degraded
quality.