Bind M-S to insert dailystatus.org in emacs
Enable auto-fill-mode when in org mode.  This gives me automatic line
wrapping, which is useful because I want wrapped lines for email.

Also bind M-S to insert the contents of ~/dailystatus.org in to the
current buffer.  This is what I do to send my daily status report.
Hopefully I can extend this in the future to construct and send the
email directly, but I am not there yet.