Add config file for xsnow
xsnow is an old X program that I like running in the winter.  The
config file turns off all the crap except the snow because I don't
care about the other stuff.  To run it, it's "xsnow -hidemenu &" and
that's it.  It will read the config file and move on.