Update mutt configuration to use w3m to display HTML email
w3m is a much better text mode browser and with a few tweaks, it's
tolerable as a tool to read HTML.  I have it display the rendered page
but with numbered links and those links listed at the bottom.

auto_view with w3m and image support does not work due to a limitation
in mutt.  It is possible to have it separately launch w3m to display
images, but you have to bind that to a macro.  I may do that, but for
now I want to see if I can get by without it.