Add 'git work' and 'git personal' command aliases
Drop the part of .zshrc that sets the git user name and email based on
the hostname.  Modify .gitconfig to set just the name and only read
the config file.  Add git aliases called work and personal to set the
appropriate email address based on the project.

Run "git work" in work projects before committing.  And run "git
personal" in non-work projects.