Remove work email account configuration for mutt
I have migrated my work email to Thunderbird because of the calendar
integration and the chat support that works with Matrix which is what
we now use for chat.fedoraproject.org.  I am keeping personal email in
mutt for now, but may migrate to Thunderbird for that as well just so
I have one email client.  But right now it's just work email.