Remove mutt and mbsync configuration files.
Well I did it.  I moved to Thunderbird after making zero progress ever
making modern day email remotely usable in mutt.  The problem is I
need to be able to both read HTML email and make use of links in those
messages.  That has just never worked for shit in mutt.  The only
thing you can do is load the message in an actual graphical browser,
which leads me to wonder what the hell the point is of mutt if I just
have to use Firefox to read all email.

I've moved back to Thunderbird as I have done in the past which is to
move back and forth between mutt and Thunderbird.  Thunderbird works
well for locally archiving email.  It also handles multiple accounts
in a clean way.  It also supports chat systems, which I have made use
of for Matrix in Fedora since we have moved there from IRC.

Will this solve all of my email problems?  Probably not, but so far it
is easier to deal with email now since everyone sends HTML email with
images and links.