Add optin() and optout() functions for /opt/*/bin PATH management
Using optin, I can add an /opt/*/bin directory to my PATH and then
undo it later with optout.  I install things to /opt with --prefix set
to /opt/PROGRAM for the purposes of testing things out or exploring
new software.  Removal is just a matter of removing that entire tree,
but to run the software I need to have that PROGRAM's bin directory in
the PATH.  These functions help with that.